refactoring syncKeys to user helper function so webhook can also use
